requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
yoomoney==0.1.1
requests==2.32.3
